third_party/WebKit/LayoutTests/fast/dom/DeviceMotion/create-event-expected.txt

CONSOLE MESSAGE: line 3: Uncaught NotSupportedError: The implementation did not support the requested type of object or operation.
Tests that document.createEvent() works with DeviceMotionEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
